+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robots/kraken/PickAndPlace.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.robots.kraken;
 
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.Servo;
